Moved the OAuth githubStrategy logic to a separate file.
